// src/app/models/ariziModels/ariziPersonel.ts
export interface AriziPersonel {
  Id: number;
  UetdsSeferReferansNo: number;
  TurKodu: number;
  UyrukUlke: string;
  TcKimlikPasaportNo: string;
  Cinsiyet: string;
  Adi: string;
  Soyadi: string;
  Telefon: string;
  Adres: string;
  HesKodu: string;
  AktifIptal: string;
  IptalAciklama: string;
}