// src/app/models/ariziModels/ariziPersonelDto.ts
export interface AriziPersonelDto {
  Id: number;
  UetdsSeferReferansNo: number;
  TurKodu: number;
  TurAdi: string;
  UyrukUlke: string;
  TcKimlikPasaportNo: string;
  Cinsiyet: string;
  Adi: string;
  Soyadi: string;
  Telefon: string;
  Adres: string;
  HesKodu: string;
  AktifIptal: string;
  IptalAciklama: string;
}